requirements.txt
Flask~=2.3.2
flask_sqlalchemy~=3.0.5
flask_bootstrap~=3.3.7.1
requests~=2.29.0
WTForms~=3.0.1
python-dotenv~=1.0.0
Flask-WTF~=1.1.1
w3lib~=2.1.2
SQLAlchemy~=1.4.49
pip~=22.2.2
Jinja2~=3.1.2
MarkupSafe~=2.1.3
Werkzeug~=2.3.6
click~=8.1.4
blinker~=1.6.2
itsdangerous~=2.1.2
visitor~=0.1.3
certifi~=2022.12.7
dominate~=2.8.0
greenlet~=2.0.2
idna~=3.4
urllib3~=1.26.15
setuptools~=68.1.2
packaging~=23.1
pytest~=7.3.2